fix compilation error: type mismatch of operands of std::max

+++ src/storage/lsh_index_storage.cpp
@@ -347,7 +347,7 @@ bool lsh_index_storage::retrieve_hit_rows(size_t hash,
   for (lsh_table_t::const_iterator it = range.first; it != range.second; ++it) {
     cands.insert(it->second);
   }
-  return cands.size() >= max(static_cast<long long unsigned>(ret_num), table_num_ * 3);
+  return cands.size() >= max(static_cast<uint64_t>(ret_num), table_num_ * 3);
 }
 
 void lsh_index_storage::get_sorted_similar_rows(const unordered_set<uint64_t>& cands,
